+++ client/magic/magicInput.jsx
@@ -1,131 +1,9 @@
 var MagicSuggestions = require("./magicSuggestions.jsx");
 var eventBus = require("../eventBus.js");
 var magic = require("./magic.js");
+var magicInputStore = require('./magicInputStore.js');
+var isKey = require('../utilities.js').isKey;
 
-
-window._keys = {
-  9: 'TAB',
-  40: 'DOWN_ARROW',
-  38: 'UP_ARROW',
-  13: 'ENTER'
-};
-
-function isKey(event){
-  var keycode = event.which;
-  var result = false;
-  for(var i=1;i<arguments.length;i++){
-    result = result || window._keys[keycode] === arguments[i];
-  }
-  return result;
-};
-function MagicInputStore (eventName){
-  var initialState = {
-      args: null,
-      suggestions: [],
-      argsSuggestions: [],
-      suggestionArgsActive: -1,
-      suggestionActive: 0,
-      preArgsLength: 0
-  };
-
-  var state = _.cloneDeep(initialState);
-  var render = function(){
-    eventBus.emit(eventName);
-  };
-  // Methods returns itself, and allows last parameter to be an optional boolean that specifies rerender.
-  var methods = {
-    getState: function(){
-      return state;
-    },
-    resetState: function(rend){
-      state = _.cloneDeep(initialState);
-      if(arguments[arguments.length-1]){
-        render();
-      }
-      return methods;
-    },
-    setActiveSuggestion: function(sug, rend){
-      state.suggestionActive = sug;
-      if(arguments[arguments.length-1]){
-        render();
-      }
-      return methods;
-    },
-    setActiveArgsSuggestion: function(sug, rend){
-      state.suggestionArgsActive = sug;
-      if(arguments[arguments.length-1]){
-        render();
-      }
-      return methods;
-    },
-    getCurrentCommand: function(){
-      return state.suggestions[state.suggestionActive];
-    },
-    getCurrentArgs: function(){
-      if(state.suggestionArgsActive < 0){
-        return state.args;
-      }else{
-        var currentArgs = state.argsSuggestions[state.suggestionArgsActive].name;
-        return currentArgs === undefined ? state.args : currentArgs;
-      }
-    },
-    setSuggestions: function(suggestions, rend){
-      state.suggestions = suggestions;
-      if(arguments[arguments.length-1]){
-        render();
-      }
-      return methods;
-    },
-    setArgsSuggestions: function(suggestions, rend){
-      state.argsSuggestions = suggestions;
-      if(arguments[arguments.length-1]){
-          render();
-      }
-      return methods;
-    },
-    setArguments: function(args){
-      state.args = args;
-      return methods;
-    },
-    activeSuggestionUp: function(rend){
-      if(state.argsSuggestions){
-        var active = (state.suggestionArgsActive - 1);
-        if(active < -1){
-          active = state.argsSuggestions.length-1;
-        }
-        methods.setActiveArgsSuggestion(active);
-      }else{
-        var active = (state.suggestionActive - 1);
-        if(active < 0){
-          active = state.suggestions.length-1;
-        }
-        methods.setActiveSuggestion(active);
-      }
-      if(arguments[arguments.length-1]){
-          render();
-      }
-      return methods;
-    },
-    activeSuggestionDown: function(rend){
-      if(state.argsSuggestions){
-        var active = (state.suggestionActive + 1) % state.suggestions.length;
-        methods.setActiveArgsSuggestion(active);
-      }else{
-        var active = (state.suggestionActive + 1) % state.suggestions.length;
-        methods.setActiveSuggestion(active);
-      }
-      if(arguments[arguments.length-1]){
-          render();
-      }
-      return methods;
-    },
-    render: render
-  };
-
-  return methods;
-};
-
-var magicInputStore = MagicInputStore('magicInput');
 var MagicInput = React.createClass({
   componentDidMount: function(){
     eventBus.register('magicInput', function() {
